Develop

+++ global_models/borrow_request_response.go
@@ -9,7 +9,7 @@ type BorrowRequestResponse struct {
 	Firstname       string     `json:"firstname"`
 	Lastname        string     `json:"lastname"`
 	RequestedAmount float64    `json:"requested_amount"`
-	RemainingAmount *float64   `json:"remaining_amount"`
+	RemainingAmount *int       `json:"remaining_amount"`
 	RequestedAt     time.Time  `json:"requested_at"`
 	DueDate         *time.Time `json:"due_date"`
 }
